# openapi.yaml
openapi: 3.0.0
info:
  title: Hospital Patient Search API
  description: |-
    REST API for searching and retrieving patient information across hospitals.
    Staff must authenticate with a Bearer JWT token to access protected endpoints.

    - 🔓 No authentication required
    - 🔐 Bearer token (JWT) authentication required
  contact:
    email: vorapon_s@hotmail.com
  version: 1.0.0
servers:
  - url: https://hossearchapi-1071770156665.asia-southeast3.run.app
    description: Production server
  - url: http://localhost:3456
    description: Local development server
...
security:
  - bearerAuth: []
paths:
...
  /patient/search/{id}:
    get:
      tags:
        - patient
      summary: Get patient by national ID or passport ID
      description: |-
        🔐 Returns a single patient's full information by their national ID (Thai 13-digit)
        or passport ID. The system determines the ID type automatically.
        Staff can only retrieve patients within their own hospital.
      operationId: getPatientById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Patient national ID (13-digit Thai) or passport ID
          schema:
            type: string
          examples:
            nationalId:
              summary: Thai national ID
              value: "1100100012341"
            passportId:
              summary: Passport ID
              value: "JP10234567"
      responses:
        '200':
          description: Patient found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientDetail'
              examples:
                thaiPatient:
                  summary: Thai patient with national ID
                  value:
                    first_name_th: "สมชาย"
                    middle_name_th: null
                    last_name_th: "ใจดี"
                    first_name_en: "Somchai"
                    middle_name_en: null
                    last_name_en: "Jaidee"
                    date_of_birth: "1985-03-15T00:00:00Z"
                    patient_hn: "BKH-0001"
                    national_id: "1100100012341"
                    passport_id: null
                    phone_number: "0812345001"
                    email: "somchai.j@email.com"
                    gender: "male"
                foreignPatientWithMiddleName:
                  summary: US patient with passport ID and middle name
                  value:
                    first_name_th: null
                    middle_name_th: null
                    last_name_th: null
                    first_name_en: "John"
                    middle_name_en: "Michael"
                    last_name_en: "Smith"
                    date_of_birth: "1982-05-10T00:00:00Z"
                    patient_hn: "BKH-0007"
                    national_id: null
                    passport_id: "US10234567"
                    phone_number: "+12125550001"
                    email: "john.m.smith@email.com"
                    gender: "male"
        '400':
          description: Bad request — ID is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "INVALID_INPUT"
                message: "Patient ID is required"
        '401':
          description: Unauthorized — missing or invalid Bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "UNAUTHORIZED"
                message: "Invalid or expired token"
        '404':
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "NOT_FOUND"
                message: "Patient not found"
        '429':
          description: Too many requests — rate limit exceeded (100 req/min per IP)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "RATE_LIMIT_EXCEEDED"
                message: "Too many requests, please try again later"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "INTERNAL_ERROR"
                message: "An unexpected error occurred"

  # ============================================================
  # STAFF / AUTH
  # ============================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PatientDetail:
      type: object
      description: Full patient record returned by the API
      properties:
        first_name_th:
          type: string
          nullable: true
          example: "สมชาย"
        middle_name_th:
          type: string
          nullable: true
          example: null
        last_name_th:
          type: string
          nullable: true
          example: "ใจดี"
        first_name_en:
          type: string
          nullable: true
          example: "Somchai"
        middle_name_en:
          type: string
          nullable: true
          example: null
        last_name_en:
          type: string
          nullable: true
          example: "Jaidee"
        date_of_birth:
          type: string
          format: date-time
          nullable: true
          description: Date of birth serialised as RFC3339 by Go's time.Time (e.g. "1985-03-15T00:00:00Z")
          example: "1985-03-15T00:00:00Z"
        patient_hn:
          type: string
          nullable: true
          description: Hospital number (unique within hospital)
          example: "BKH-0001"
        national_id:
          type: string
          nullable: true
          description: Thai national ID (13 digits)
          example: "1100100012341"
        passport_id:
          type: string
          nullable: true
          description: Passport number
          example: null
        phone_number:
          type: string
          nullable: true
          example: "0812345001"
        email:
          type: string
          nullable: true
          example: "somchai.j@email.com"
        gender:
          type: string
          nullable: true
          enum: [male, female, other]
          example: "male"
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: "NOT_FOUND"
        message:
          type: string
          description: Human-readable error description
          example: "The requested resource was not found"